Created seperate controller for tasks page
